test(spdx): Extract the test input `OrtResult` to `testFixtures`
Prepare for sharing this test data with the tests of an upcoming new
implementation of an SPDX reporter, in context of [1].

[1]: https://github.com/oss-review-toolkit/ort/issues/9998
